requirements.txt
datasets
transformers
soundfile
librosa
torch
torchvision 
torchaudio
webrtcvad
onnx
onnxruntime
pyctcdecode
pyaudio
sentencepiece
openai
python-dotenv
streamlit
llama-cpp-python